refactor: Move dashboard head to a separate component
- And move it above dashboard content so that it takes up the top space and dashboard content doesn't partial hide under the head on scroll
